h2: hpack: C-020: dynamic table hdr_len must not truncate
A literal header name can run to the whole header block, but the
dynamic table entry stored hdr_len as uint16 while the full width was
added to virtual_payload_usage at insert.  Eviction then subtracted the
truncated value, so usage drifted upwards by 64KB per such entry until
every insert evicted the whole table and the peer's next indexed
reference failed with COMPRESSION_ERROR.  Widen the field and subtract
exactly what was added.

+++ lib/roles/h2/hpack.c
@@ -472,7 +472,7 @@ lws_h2_dynamic_table_dump(struct lws *wsi)
 			p = "(ignored)";
 		lwsl_header("   %3d: tok %s: (len %d) val '%s'\n",
 			    (int)(n + LWS_ARRAY_SIZE(static_token)), p,
-			    dyn->entries[m].hdr_len, dyn->entries[m].value ?
+			    (int)dyn->entries[m].hdr_len, dyn->entries[m].value ?
 			    dyn->entries[m].value : "null");
 	}
 #endif
@@ -483,8 +483,10 @@ static void
 lws_dynamic_free(struct hpack_dynamic_table *dyn, int idx)
 {
 	lwsl_header("freeing %d for reuse\n", idx);
-	dyn->virtual_payload_usage = (uint32_t)((unsigned int)dyn->virtual_payload_usage - (unsigned int)(dyn->entries[idx].value_len +
-				dyn->entries[idx].hdr_len));
+	/* must subtract exactly what lws_dynamic_token_insert() added */
+	dyn->virtual_payload_usage = (uint32_t)(dyn->virtual_payload_usage -
+				((uint32_t)dyn->entries[idx].value_len +
+				 dyn->entries[idx].hdr_len));
 	lws_free_set_NULL(dyn->entries[idx].value);
 	dyn->entries[idx].value = NULL;
 	dyn->entries[idx].value_len = 0;
@@ -582,7 +584,7 @@ lws_dynamic_token_insert(struct lws *wsi, int hdr_len,
 		dyn->entries[new_index].value = NULL;
 
 	dyn->entries[new_index].lws_hdr_idx = (uint16_t)lws_hdr_index;
-	dyn->entries[new_index].hdr_len = (uint16_t)hdr_len;
+	dyn->entries[new_index].hdr_len = (uint32_t)hdr_len;
 
 	dyn->virtual_payload_usage = (uint32_t)(dyn->virtual_payload_usage +
 					(unsigned int)hdr_len + len);
